Adding attribute usage to Echo files
